+++ HISTORY.rst
@@ -5,6 +5,9 @@ History
 1.0.4dev
 --------
 
+- add ``novalidate`` property for form
+  [rnix, 2011-05-23]
+
 - return empty string in mode renderer if value is UNSET
   [rnix, 2011-05-23]